refactor(platform): change api validation logic
